fix(erp): fix all Transactions workflow issues for release
Issue 1 & 2 — Approve/Reject type mismatch (inconsistent types for $2):
  reviewed_by is BIGINT (base schema) but approved_by/rejected_by are INT
  (migration 065). Using $2 for both caused PostgreSQL parameter type
  deduction error. Fixed by casting to $2::int on all new INT columns.

Issue 1 extra — Reject routes had no try/catch: unhandled exceptions
  returned 500 HTML which the frontend couldn't parse, showing the default
  "Failed to reject" message. Added proper try/catch with JSON error responses
  to both deposits/reject and withdrawals/reject.

Issue 3 — Approve/Reject buttons not visible (fixed in prev commit 665a520):
  /api/auth/me returns `sub` not `id`; Handle page now reads d.sub correctly.

Issue 4 — Deposit Bank info incomplete (fixed in prev commit 665a520):
  COALESCE JOIN fallback now looks up payment_banks by bank_name when
  receiving_bank_id is NULL (legacy deposits).

Trigger fix — Migration 063 trigger only handled PENDING → PAID/REJECTED.
  With PROCESSING status added in 065, PROCESSING → PAID/REJECTED transitions
  did not decrement pending_withdrawal. Fixed via migration 066 + inline
  migrate route step.

Migration route — Added migrations 065 and 066 to /api/settings/migrate
  POST so they can be applied via the ERP Maintenance → Run Migrations button.

+++ erp/src/app/api/deposits/[id]/approve/route.ts
@@ -41,7 +41,7 @@ export async function POST(
     await client.query(
       `UPDATE deposit_requests
        SET status = 'APPROVED', reviewed_by = $2, admin_note = $3, reviewed_at = NOW(),
-           approved_by = $2, approved_at = NOW()
+           approved_by = $2::int, approved_at = NOW()
        WHERE id = $1`,
       [requestId, adminId, null]
     );
