+++ erp/src/app/api/withdrawals/[id]/approve/route.ts
@@ -51,7 +51,7 @@ export async function POST(
     await client.query(
       `UPDATE withdrawal_requests
        SET status = 'PAID', reviewed_by = $2, admin_note = $3, reviewed_at = NOW(),
-           approved_by = $2, approved_at = NOW()
+           approved_by = $2::int, approved_at = NOW()
        WHERE id = $1`,
       [requestId, adminId, null]
     );
